+++ src/main/java/com/carportal/CarPortalApplication.java
@@ -1,0 +1,12 @@
+package com.carportal;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CarPortalApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(CarPortalApplication.class, args);
+  }
+}
