+++ src/main/java/com/carportal/audit/Auditable.java
@@ -1,13 +1,10 @@
 package com.carportal.audit;
 
-import static javax.persistence.TemporalType.TIMESTAMP;
-
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import jakarta.persistence.Temporal;
+import jakarta.persistence.TemporalType;
 import java.util.Date;
-
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import javax.persistence.Temporal;
-
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
@@ -16,28 +13,28 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 @EntityListeners(AuditingEntityListener.class)
 public abstract class Auditable<T> {
 
-	@CreatedDate
-	@Temporal(TIMESTAMP)
-	protected Date createdDate;
+  @CreatedDate
+  @Temporal(TemporalType.TIMESTAMP)
+  protected Date createdDate;
 
-	@LastModifiedDate
-	@Temporal(TIMESTAMP)
-	protected Date lastModifiedDate;
+  @LastModifiedDate
+  @Temporal(TemporalType.TIMESTAMP)
+  protected Date lastModifiedDate;
 
-	public Date getCreatedDate() {
-		return createdDate;
-	}
+  public Date getCreatedDate() {
+    return createdDate;
+  }
 
-	public void setCreatedDate(Date createdDate) {
-		this.createdDate = createdDate;
-	}
+  public void setCreatedDate(Date createdDate) {
+    this.createdDate = createdDate;
+  }
 
-	public Date getLastModifiedDate() {
-		return lastModifiedDate;
-	}
+  public Date getLastModifiedDate() {
+    return lastModifiedDate;
+  }
 
-	public void setLastModifiedDate(Date lastModifiedDate) {
-		this.lastModifiedDate = lastModifiedDate;
-	}
+  public void setLastModifiedDate(Date lastModifiedDate) {
+    this.lastModifiedDate = lastModifiedDate;
+  }
 
 }
