+++ src/main/java/com/carportal/config/CustomGeneratorRegistrar.java
@@ -1,0 +1,14 @@
+package com.carportal.config;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class CustomGeneratorRegistrar implements GenerationTypeStrategyRegistrationProvider {
+
+  @Override
+  public Map<String, Class<?>> getRegisteredGenerators() {
+    Map<String, Class<?>> map = new HashMap<>();
+    map.put("prefixed-sequence", PrefixedSequenceIdGenerator.class);
+    return map;
+  }
+}
