--- src/main/java/com/carportal/config/MyWebAppInitializer.java
@@ -1,30 +1,0 @@
-package com.carportal.config;
-
-import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRegistration;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.web.WebApplicationInitializer;
-import org.springframework.web.context.ContextLoaderListener;
-import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
-import org.springframework.web.servlet.DispatcherServlet;
-
-@ComponentScan(basePackages = { "com.carportal" })
-public class MyWebAppInitializer implements WebApplicationInitializer {
-
-	private static final String DISPATCHER_SERVLET_NAME = "dispatcher";
-
-	@Override
-	public void onStartup(ServletContext container) throws ServletException {
-
-		AnnotationConfigWebApplicationContext context = new AnnotationConfigWebApplicationContext();
-		context.setConfigLocation("com.carportal");
-		container.addListener(new ContextLoaderListener(context));
-		ServletRegistration.Dynamic dispatcher = container.addServlet(DISPATCHER_SERVLET_NAME,
-				new DispatcherServlet(context));
-		dispatcher.setLoadOnStartup(1);
-		dispatcher.addMapping("/");
-
-	}
-}
