+++ src/main/java/com/carportal/entity/CarEngine.java
@@ -1,0 +1,59 @@
+package com.carportal.entity;
+
+import com.carportal.constants.ApplicationConstants.Db;
+import com.fasterxml.jackson.annotation.JsonIdentityInfo;
+import com.fasterxml.jackson.annotation.ObjectIdGenerators;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.MapsId;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(schema = Db.TBL_SCHEMA_CAR_PORTAL, name = Db.TBL_CAR_ENGINE)
+@JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class, property = "carEngineId", scope = Long.class)
+public class CarEngine {
+
+  @Id
+  private Long carEngineId;
+
+  @Column(nullable = false)
+  private double mileage;
+
+  @Column(nullable = false)
+  private String transmissionType;
+
+  @Column(nullable = false)
+  private String engineType;
+
+  @Column(nullable = false)
+  private double engineDisplacement;
+
+  @Column(nullable = false)
+  private double maxPower;
+
+  @Column(nullable = false)
+  private double maxTourque;
+
+  @Column(nullable = false)
+  private int numberOfCylinder;
+
+  @Column(nullable = false)
+  private String fuel;
+
+  @Column(nullable = false)
+  private String driveType;
+
+  @MapsId
+  @OneToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "carDetailId")
+  private CarDetails carDetails;
+
+}
