+++ src/main/java/com/carportal/mapper/CarEngineMapper.java
@@ -1,0 +1,9 @@
+package com.carportal.mapper;
+
+import com.carportal.entity.CarEngine;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring", uses = CarEngine.class)
+public interface CarEngineMapper {
+
+}
