+++ src/main/java/com/carportal/mapper/CarPhotoMapper.java
@@ -1,0 +1,9 @@
+package com.carportal.mapper;
+
+import com.carportal.entity.CarPhoto;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring", uses = CarPhoto.class)
+public interface CarPhotoMapper {
+
+}
