--- src/main/java/com/carportal/repository/CarDetailRepositoryCustom.java
@@ -1,5 +1,0 @@
-package com.carportal.repository;
-
-public interface CarDetailRepositoryCustom {
-
-}
