+++ src/main/java/com/carportal/repository/CarEngineRepository.java
@@ -1,9 +1,8 @@
 package com.carportal.repository;
 
+import com.carportal.entity.CarEngine;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import com.carportal.model.CarEngine;
 
 @Repository
 public interface CarEngineRepository extends JpaRepository<CarEngine, Long> {
