+++ src/main/java/com/carportal/repository/CarFeaturesRepository.java
@@ -1,8 +1,7 @@
 package com.carportal.repository;
 
+import com.carportal.entity.CarFeatures;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.carportal.model.CarFeatures;
 
 public interface CarFeaturesRepository extends JpaRepository<CarFeatures, Long> {
