+++ src/main/java/com/carportal/repository/CarOuterRepository.java
@@ -1,8 +1,7 @@
 package com.carportal.repository;
 
+import com.carportal.entity.CarOuter;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.carportal.model.CarOuter;
 
 public interface CarOuterRepository extends JpaRepository<CarOuter, Long> {
