+++ src/main/java/com/carportal/repository/CarPhotoRepository.java
@@ -1,8 +1,7 @@
 package com.carportal.repository;
 
+import com.carportal.entity.CarPhoto;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.carportal.model.CarPhoto;
 
 public interface CarPhotoRepository extends JpaRepository<CarPhoto, Long> {
