+++ src/main/java/com/carportal/service/CarEngineService.java
@@ -1,20 +1,18 @@
 package com.carportal.service;
 
+import com.carportal.entity.CarEngine;
 import java.util.Optional;
-
 import org.springframework.stereotype.Service;
-
-import com.carportal.model.CarEngine;
 
 @Service
 public interface CarEngineService {
 
-	public abstract void saveCarEngine(CarEngine entity, Long carDetailId);
+  void saveCarEngine(CarEngine entity, Long carDetailId);
 
-	public abstract Optional<CarEngine> findById(Long carEngineId);
+  Optional<CarEngine> findById(Long carEngineId);
 
-	public abstract Optional<CarEngine> updateById(CarEngine carEngine, Long carEngineId);
+  Optional<CarEngine> updateById(CarEngine carEngine, Long carEngineId);
 
-	public abstract void deleteById(Long carDetailId);
+  void deleteById(Long carDetailId);
 
 }
