+++ src/main/java/com/carportal/service/CarOuterService.java
@@ -1,17 +1,16 @@
 package com.carportal.service;
 
+import com.carportal.entity.CarOuter;
 import java.util.Optional;
-
-import com.carportal.model.CarOuter;
 
 public interface CarOuterService {
 
-	public abstract void saveCarOuter(CarOuter entity, Long carDetailId);
+  void saveCarOuter(CarOuter entity, Long carDetailId);
 
-	public abstract Optional<CarOuter> findById(Long carEngineId);
+  Optional<CarOuter> findById(Long carEngineId);
 
-	public abstract Optional<CarOuter> updateById(CarOuter carOuter, Long carOuterId);
+  Optional<CarOuter> updateById(CarOuter carOuter, Long carOuterId);
 
-	public abstract void deleteById(Long carDetailId);
+  void deleteById(Long carDetailId);
 
 }
