+++ src/main/resources/application.properties
@@ -1,0 +1,17 @@
+# Database
+spring.datasource.url=jdbc:mysql://localhost:3306/carportal
+spring.datasource.username=root
+spring.datasource.password=your_password
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+# Hibernate / JPA
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
+spring.jpa.properties.hibernate.format_sql=true
+# Hibernate cache (optional)
+spring.jpa.properties.hibernate.cache.use_second_level_cache=false
+spring.jpa.properties.hibernate.cache.use_query_cache=false
+# Connection pool tuning (optional)
+spring.datasource.hikari.maximum-pool-size=16
+spring.datasource.hikari.auto-commit=false
+spring.jpa.properties.hibernate.generator.identifier_generator_strategy_provider=com.carportal.config.CustomGeneratorRegistrar
