+++ src/main/resources/logback.xml
@@ -1,56 +1,40 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<configuration debug="true">
+<configuration debug="false">
 
-	<property name="HOME_LOG" value="logs/app.log" />
+  <property name="LOG_DIR" value="${LOG_DIR:-./logs}"/>
 
-	<appender name="CONSOLE"
-		class="ch.qos.logback.core.ConsoleAppender">
-		<layout class="ch.qos.logback.classic.PatternLayout">
-			<Pattern>
-				%date{HH:mm:ss.SSS} [%thread] [%-5level] %logger{40} - %message%n
-			</Pattern>
-		</layout>
-	</appender>
+  <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
+    <layout class="ch.qos.logback.classic.PatternLayout">
+      <pattern>%d{HH:mm:ss.SSS} [%thread] [%-5level] %logger{40} - %message%n</pattern>
+    </layout>
+  </appender>
 
-	<appender name="FILE-ROLLING"
-		class="ch.qos.logback.core.rolling.RollingFileAppender">
-		<file>${HOME_LOG}</file>
+  <appender name="FILE-ROLLING" class="ch.qos.logback.core.rolling.RollingFileAppender">
+    <file>${LOG_DIR}/app.log</file>
+    <rollingPolicy class="ch.qos.logback.core.rolling.SizeAndTimeBasedRollingPolicy">
+      <fileNamePattern>${LOG_DIR}/archived/app.%d{yyyy-MM-dd}.%i.log</fileNamePattern>
+      <maxFileSize>10MB</maxFileSize>
+      <totalSizeCap>20GB</totalSizeCap>
+      <maxHistory>60</maxHistory>
+    </rollingPolicy>
+    <encoder>
+      <pattern>%d{yyyy-MM-dd HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+    </encoder>
+  </appender>
 
-		<rollingPolicy
-			class="ch.qos.logback.core.rolling.SizeAndTimeBasedRollingPolicy">
-			<fileNamePattern>logs/archived/app.%d{yyyy-MM-dd}.%i.log
-			</fileNamePattern>
-			<!-- each archived file, size max 10MB -->
-			<maxFileSize>10MB</maxFileSize>
-			<!-- total size of all archive files, if total size > 20GB, it will delete 
-				old archived file -->
-			<totalSizeCap>20GB</totalSizeCap>
-			<!-- 60 days to keep -->
-			<maxHistory>60</maxHistory>
-		</rollingPolicy>
+  <appender name="ASYNC" class="ch.qos.logback.classic.AsyncAppender">
+    <queueSize>512</queueSize>
+    <appender-ref ref="FILE-ROLLING"/>
+  </appender>
 
-		<encoder>
-			<pattern>%d %p %c{1.} [%t] %m%n</pattern>
-		</encoder>
-	</appender>
+  <logger name="com.carportal" level="info" additivity="false">
+    <appender-ref ref="ASYNC"/>
+    <appender-ref ref="CONSOLE"/>
+  </logger>
 
-	<!-- https://logback.qos.ch/manual/appenders.html#AsyncAppender -->
-	<!-- AsyncAppender will drop events of level TRACE, DEBUG and INFO if its 
-		queue is 80% full -->
-	<appender name="ASYNC"
-		class="ch.qos.logback.classic.AsyncAppender">
-		<appender-ref ref="FILE-ROLLING" />
-		<!-- defaulr 256 -->
-		<queueSize>512</queueSize>
-	</appender>
-
-	<logger name="com.carportal" level="info" additivity="false">
-		<appender-ref ref="ASYNC" />
-		<appender-ref ref="CONSOLE" />
-	</logger>
-
-	<root level="info">
-		<appender-ref ref="CONSOLE" />
-	</root>
+  <root level="info">
+    <appender-ref ref="CONSOLE"/>
+    <appender-ref ref="ASYNC"/>
+  </root>
 
 </configuration>
